changed status and removed devices
